feat: 땅입력 처리 방식 수정
- 니어리스트 이후 다 닫고, 땅 팝업 작동하게 방식 변경
  -> 땅 팝업에서 취소 누를 경우 기존에는 다음홀 처리 되었으나 취소 시, 미입력전으로 갈 수 있음
- TODO : 땅팝업 리턴 타입 관련 수정 필요
